exception handled partially

+++ rms/src/main/java/com/rms/service/impl/CategoryServiceImpl.java
@@ -7,6 +7,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import com.rms.constants.ApplicationConstants;
 import com.rms.dao.CategoryDao;
@@ -14,6 +15,8 @@ import com.rms.dto.CategoryDto;
 import com.rms.entity.Category;
 import com.rms.exception.BusinessLogicException;
 import com.rms.exception.DataBaseException;
+import com.rms.exception.IdNotFoundException;
+import com.rms.exception.NoRecordFoundException;
 import com.rms.service.CategoryService;
 import com.rms.util.CategoryUtil;
 
@@ -30,7 +33,13 @@ public class CategoryServiceImpl implements CategoryService {
 	public String deleteCategory(Long id) {
 		logger.info("Entering deleteCategory method");
 		try {
-			return categoryDao.deleteCategory(id);
+			if(categoryDao.getCategoryById(id)!=null) {
+				return categoryDao.deleteCategory(id);
+			}
+			else {
+				throw new IdNotFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
+			}
+			
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
 			throw new BusinessLogicException(e.getMessage());
@@ -41,6 +50,7 @@ public class CategoryServiceImpl implements CategoryService {
 	public String updateCategory(Long id, CategoryDto categoryDto) {
 		logger.info("Entering updateCategory method");
 		try {
+			if(categoryDao.getCategoryById(id)!=null) {
 			String result = null;
 			Category category = CategoryUtil.toEntity(categoryDto);
 			boolean flag = categoryDao.updateCategory(id, category);
@@ -48,6 +58,11 @@ public class CategoryServiceImpl implements CategoryService {
 				result = ApplicationConstants.CATEGORY_UPDATE_SUCCESS;
 			}
 			return result;
+			}
+			else {
+				throw new IdNotFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
+			}
+			
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
 			throw new BusinessLogicException(e.getMessage());
@@ -78,8 +93,8 @@ public class CategoryServiceImpl implements CategoryService {
 			Category category = categoryDao.getCategoryById(id);
 			if (category != null) {
 				return CategoryUtil.toDto(category);
-			} else {
-				throw new BusinessLogicException(ApplicationConstants.CATEGORY_NOT_FOUND);
+			}else {
+				throw new NoRecordFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -92,12 +107,13 @@ public class CategoryServiceImpl implements CategoryService {
 		logger.info("Entering getAllCategory method");
 		try {
 			List<Category> categoryEntity = categoryDao.getAllCategory();
-			if (categoryEntity != null) {
+			if (CollectionUtils.isEmpty(categoryEntity)) {
+				throw new NoRecordFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
+			}else {
+			
 				List<CategoryDto> categoryDto = new ArrayList<>();
 				categoryEntity.stream().forEach(entity -> categoryDto.add(CategoryUtil.toDto(entity)));
 				return categoryDto;
-			} else {
-				throw new BusinessLogicException(ApplicationConstants.CATEGORY_NOT_FOUND);
 			}
 
 		} catch (DataBaseException e) {
