+++ rms/src/main/java/com/rms/service/impl/ProductServiceImpl.java
@@ -1,5 +1,6 @@
 package com.rms.service.impl;
 
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -7,6 +8,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import com.rms.constants.ApplicationConstants;
 import com.rms.dao.CategoryDao;
@@ -18,8 +20,11 @@ import com.rms.entity.Meal;
 import com.rms.entity.Product;
 import com.rms.exception.BusinessLogicException;
 import com.rms.exception.DataBaseException;
+import com.rms.exception.IdNotFoundException;
+import com.rms.exception.NoRecordFoundException;
 import com.rms.service.ProductService;
 import com.rms.util.ProductUtil;
+import com.rms.util.TimeConverterUtil;
 
 @Service
 public class ProductServiceImpl implements ProductService {
@@ -39,8 +44,13 @@ public class ProductServiceImpl implements ProductService {
 	public String deleteProduct(Long id) {
 		logger.info("Entering deleteProduct method");
 		try {
+			if(productDao.getProductById(id)!=null) {
 			Product product = productDao.getProductById(id);
 			return productDao.deleteProduct(product);
+			}
+			else {
+				throw new IdNotFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
+			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
 			throw new BusinessLogicException(e.getMessage());
@@ -63,10 +73,10 @@ public class ProductServiceImpl implements ProductService {
 					}
 					return result;
 				} else {
-					throw new BusinessLogicException(ApplicationConstants.MEAL_NOT_FOUND);
+					throw new NoRecordFoundException(ApplicationConstants.MEAL_NOT_FOUND);
 				}
 			} else {
-				throw new BusinessLogicException(ApplicationConstants.CATEGORY_NOT_FOUND);
+				throw new NoRecordFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -92,10 +102,10 @@ public class ProductServiceImpl implements ProductService {
 					}
 					return result;
 				} else {
-					throw new BusinessLogicException(ApplicationConstants.MEAL_NOT_FOUND);
+					throw new NoRecordFoundException(ApplicationConstants.MEAL_NOT_FOUND);
 				}
 			} else {
-				throw new BusinessLogicException(ApplicationConstants.CATEGORY_NOT_FOUND);
+				throw new NoRecordFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -112,7 +122,7 @@ public class ProductServiceImpl implements ProductService {
 			if (product != null) {
 				return ProductUtil.toDto(product);
 			} else {
-				throw new BusinessLogicException(ApplicationConstants.PRODUCT_NOT_FOUND);
+				throw new IdNotFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -125,12 +135,13 @@ public class ProductServiceImpl implements ProductService {
 		logger.info("Entering getAllProduct method");
 		try {
 			List<Product> productEntity = productDao.getAllProduct();
-			if (productEntity != null) {
+			if (CollectionUtils.isEmpty(productEntity)) {
+				throw new NoRecordFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
+			}
+			else {
 				List<ProductDto> productDto = new ArrayList<>();
 				productEntity.stream().forEach(entity -> productDto.add(ProductUtil.toDto(entity)));
 				return productDto;
-			} else {
-				throw new BusinessLogicException(ApplicationConstants.PRODUCT_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -142,13 +153,19 @@ public class ProductServiceImpl implements ProductService {
 	public List<ProductDto> getProductByTypeAndCategory(Long categoryId, String type) {
 		logger.info("Entering getProductByTypeAndCategory method");
 		try {
+			
+			if(categoryDao.getCategoryById(categoryId)==null) {
+				throw new IdNotFoundException(ApplicationConstants.CATEGORY_NOT_FOUND);
+			}
+			
 			List<Product> productEntity = productDao.getProductByTypeAndCategory(categoryId, type);
-			if (productEntity != null) {
+			if (CollectionUtils.isEmpty(productEntity)) {
+				throw new NoRecordFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
+			}
+			else {
 				List<ProductDto> productDto = new ArrayList<>();
-				productEntity.stream().forEach(entity -> productDto.add(ProductUtil.toDto(entity)));
+				productEntity.stream().filter(p->p.getMeal().getName().equals(TimeConverterUtil.getTime(LocalTime.now()))).forEach(entity -> productDto.add(ProductUtil.toDto(entity)));
 				return productDto;
-			} else {
-				throw new BusinessLogicException(ApplicationConstants.PRODUCT_NOT_FOUND);
 			}
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
@@ -160,14 +177,39 @@ public class ProductServiceImpl implements ProductService {
 	public List<ProductDto> getProductByMeal(Long mealId) {
 		logger.info("Entering getProductByTypeAndCategory method");
 		try {
+			
+			if(mealDao.getMealById(mealId)==null) {
+				throw new IdNotFoundException(ApplicationConstants.MEAL_NOT_FOUND);
+			}
+			
 			List<Product> productEntity = productDao.getProductByMeal(mealId);
-			if (productEntity != null) {
+			if (CollectionUtils.isEmpty(productEntity)) {
+				throw new NoRecordFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
+			}
+			else {
 				List<ProductDto> productDto = new ArrayList<>();
 				productEntity.stream().forEach(entity -> productDto.add(ProductUtil.toDto(entity)));
 				return productDto;
-			} else {
-				throw new BusinessLogicException(ApplicationConstants.PRODUCT_NOT_FOUND);
 			}
+		} catch (DataBaseException e) {
+			logger.error(e.getMessage());
+			throw new BusinessLogicException(e.getMessage());
+		}
+	}
+
+	@Override
+	public List<ProductDto> getAllProductByMeal() {
+		logger.info("Entering getAllProductByMeal method");
+		try {
+			List<Product> productEntity = productDao.getAllProduct();
+			if (CollectionUtils.isEmpty(productEntity)) {
+				throw new NoRecordFoundException(ApplicationConstants.PRODUCT_NOT_FOUND);
+			}
+			else {
+				List<ProductDto> productDto = new ArrayList<>();
+				productEntity.stream().filter(p->p.getMeal().getName().equals(TimeConverterUtil.getTime(LocalTime.now()))).forEach(entity -> productDto.add(ProductUtil.toDto(entity)));
+				return productDto;
+			} 
 		} catch (DataBaseException e) {
 			logger.error(e.getMessage());
 			throw new BusinessLogicException(e.getMessage());
